Cambios

+++ Programacion3 java/EjerciciosHerencia/src/main/java/Ejericios/Herencia/entity/Sueldos/BonoA.java
@@ -3,7 +3,7 @@ package Ejericios.Herencia.entity.Sueldos;
 public class BonoA extends Presentismo{
 
     @Override
-    public double calcularBono(Integer ausentes) {
+    public double calcularBono(double ausentes) {
         if(ausentes==0)
         {
             return 1000;
