+++ Programacion3 java/EjerciciosHerencia/src/main/java/Ejericios/Herencia/entity/Sueldos/BonoB.java
@@ -3,7 +3,7 @@ package Ejericios.Herencia.entity.Sueldos;
 public class BonoB extends Presentismo{
 
     @Override
-    public double calcularBono(Integer ausentes) {
+    public double calcularBono(double ausentes) {
         return 500;
     }
 }
